refactor: routing logic into CheckoutRoute component
I found that I had to mock next router in my test which turned out to be
really difficult. In addition, this ended up becoming a test based on
implementation rather than behaviours.
